Rename InContextCoreDetails to InContextMetadata
