+++ Sources/InContextMetadata/src/metadata.c
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#include <InContextCoreDetails.h>
+#include <InContextMetadata.h>
 
 #ifndef INCONTEXT_VERSION
 #define INCONTEXT_VERSION "Unknown"
